+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/shizuku/ShizukuServiceConnection.aidl
@@ -1,6 +1,6 @@
 package eu.darken.sdmse.common.shizuku;
 
-import eu.darken.sdmse.common.files.local.root.FileOpsConnection;
+import eu.darken.sdmse.common.files.local.ipc.FileOpsConnection;
 import eu.darken.sdmse.common.pkgs.pkgops.root.PkgOpsConnection;
 import eu.darken.sdmse.common.shell.root.ShellOpsConnection;
