+++ app-common-io/src/main/java/eu/darken/sdmse/common/files/local/ipc/LocalPathLookupExtendedPayload.kt
@@ -1,21 +1,22 @@
-package eu.darken.sdmse.common.files.local.root
+package eu.darken.sdmse.common.files.local.ipc
 
 import android.os.Parcel
 import android.os.Parcelable
 import eu.darken.sdmse.common.debug.Bugs
 import eu.darken.sdmse.common.debug.logging.Logging.Priority.VERBOSE
 import eu.darken.sdmse.common.debug.logging.log
-import eu.darken.sdmse.common.files.local.LocalPathLookup
+import eu.darken.sdmse.common.files.local.LocalPathLookupExtended
 import eu.darken.sdmse.common.files.remoteInputStream
 import eu.darken.sdmse.common.ipc.RemoteInputStream
 import eu.darken.sdmse.common.ipc.inputStream
 import okio.Buffer
 
-data class LocalPathLookupsPayload(
-    val payload: List<LocalPathLookup>,
+data class LocalPathLookupExtendedPayload(
+    val payload: List<LocalPathLookupExtended>,
 ) : Parcelable {
     constructor(parcel: Parcel) : this(
-        parcel.readParcelableArray(LocalPathLookup::class.java.classLoader)!!.toList() as List<LocalPathLookup>
+        parcel.readParcelableArray(LocalPathLookupExtended::class.java.classLoader)!!
+            .toList() as List<LocalPathLookupExtended>
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
@@ -24,17 +25,17 @@ data class LocalPathLookupsPayload(
 
     override fun describeContents(): Int = 0
 
-    companion object CREATOR : Parcelable.Creator<LocalPathLookupsPayload> {
-        override fun createFromParcel(parcel: Parcel): LocalPathLookupsPayload {
+    companion object CREATOR : Parcelable.Creator<LocalPathLookupExtendedPayload> {
+        override fun createFromParcel(parcel: Parcel): LocalPathLookupExtendedPayload {
 
-            return LocalPathLookupsPayload(parcel)
+            return LocalPathLookupExtendedPayload(parcel)
         }
 
-        override fun newArray(size: Int): Array<LocalPathLookupsPayload?> = arrayOfNulls(size)
+        override fun newArray(size: Int): Array<LocalPathLookupExtendedPayload?> = arrayOfNulls(size)
     }
 }
 
-fun LocalPathLookupsPayload.toRemoteInputStream(): RemoteInputStream {
+fun LocalPathLookupExtendedPayload.toRemoteInputStream(): RemoteInputStream {
     val buffer = Buffer()
     val parcel = Parcel.obtain()
     this.writeToParcel(parcel, 0)
@@ -45,17 +46,17 @@ fun LocalPathLookupsPayload.toRemoteInputStream(): RemoteInputStream {
     return buffer.remoteInputStream()
 }
 
-fun RemoteInputStream.toLocalPathLookupsPayload(): LocalPathLookupsPayload {
+fun RemoteInputStream.toLocalPathLookupExtendedPayload(): LocalPathLookupExtendedPayload {
     val output = this.inputStream().readBytes()
     val parcel = Parcel.obtain()
     if (Bugs.isTrace) log(FileOpsClient.TAG, VERBOSE) { "Unmarshalling parcel now..." }
     parcel.unmarshall(output, 0, output.size)
     parcel.setDataPosition(0)
-    val payload = LocalPathLookupsPayload.createFromParcel(parcel)
+    val payload = LocalPathLookupExtendedPayload.createFromParcel(parcel)
     if (Bugs.isTrace) log(FileOpsClient.TAG, VERBOSE) { "Parcel unmarshalled, creating payload class..." }
     return payload
 }
 
-fun List<LocalPathLookup>.toRemoteInputStream() = LocalPathLookupsPayload(this).toRemoteInputStream()
+fun List<LocalPathLookupExtended>.toRemoteInputStream() = LocalPathLookupExtendedPayload(this).toRemoteInputStream()
 
-fun RemoteInputStream.toLocalPathLookups() = toLocalPathLookupsPayload().payload
+fun RemoteInputStream.toLocalPathLookupExtended() = toLocalPathLookupExtendedPayload().payload
