+++ app-common-io/src/main/java/eu/darken/sdmse/common/files/local/ipc/LocalPathLookupPayload.kt
@@ -1,22 +1,21 @@
-package eu.darken.sdmse.common.files.local.root
+package eu.darken.sdmse.common.files.local.ipc
 
 import android.os.Parcel
 import android.os.Parcelable
 import eu.darken.sdmse.common.debug.Bugs
 import eu.darken.sdmse.common.debug.logging.Logging.Priority.VERBOSE
 import eu.darken.sdmse.common.debug.logging.log
-import eu.darken.sdmse.common.files.local.LocalPath
 import eu.darken.sdmse.common.files.local.LocalPathLookup
 import eu.darken.sdmse.common.files.remoteInputStream
 import eu.darken.sdmse.common.ipc.RemoteInputStream
 import eu.darken.sdmse.common.ipc.inputStream
 import okio.Buffer
 
-data class LocalPathsPayload(
-    val payload: List<LocalPath>,
+data class LocalPathLookupsPayload(
+    val payload: List<LocalPathLookup>,
 ) : Parcelable {
     constructor(parcel: Parcel) : this(
-        parcel.readParcelableArray(LocalPathLookup::class.java.classLoader)!!.toList() as List<LocalPath>
+        parcel.readParcelableArray(LocalPathLookup::class.java.classLoader)!!.toList() as List<LocalPathLookup>
     )
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
@@ -25,16 +24,17 @@ data class LocalPathsPayload(
 
     override fun describeContents(): Int = 0
 
-    companion object CREATOR : Parcelable.Creator<LocalPathsPayload> {
-        override fun createFromParcel(parcel: Parcel): LocalPathsPayload {
-            return LocalPathsPayload(parcel)
+    companion object CREATOR : Parcelable.Creator<LocalPathLookupsPayload> {
+        override fun createFromParcel(parcel: Parcel): LocalPathLookupsPayload {
+
+            return LocalPathLookupsPayload(parcel)
         }
 
-        override fun newArray(size: Int): Array<LocalPathsPayload?> = arrayOfNulls(size)
+        override fun newArray(size: Int): Array<LocalPathLookupsPayload?> = arrayOfNulls(size)
     }
 }
 
-fun LocalPathsPayload.toRemoteInputStream(): RemoteInputStream {
+fun LocalPathLookupsPayload.toRemoteInputStream(): RemoteInputStream {
     val buffer = Buffer()
     val parcel = Parcel.obtain()
     this.writeToParcel(parcel, 0)
@@ -45,18 +45,17 @@ fun LocalPathsPayload.toRemoteInputStream(): RemoteInputStream {
     return buffer.remoteInputStream()
 }
 
-fun RemoteInputStream.toLocalPathsPayload(): LocalPathsPayload {
+fun RemoteInputStream.toLocalPathLookupsPayload(): LocalPathLookupsPayload {
     val output = this.inputStream().readBytes()
     val parcel = Parcel.obtain()
     if (Bugs.isTrace) log(FileOpsClient.TAG, VERBOSE) { "Unmarshalling parcel now..." }
     parcel.unmarshall(output, 0, output.size)
     parcel.setDataPosition(0)
+    val payload = LocalPathLookupsPayload.createFromParcel(parcel)
     if (Bugs.isTrace) log(FileOpsClient.TAG, VERBOSE) { "Parcel unmarshalled, creating payload class..." }
-    val payload = LocalPathsPayload.createFromParcel(parcel)
-    parcel.recycle()
     return payload
 }
 
-fun List<LocalPath>.toRemoteInputStream() = LocalPathsPayload(this).toRemoteInputStream()
+fun List<LocalPathLookup>.toRemoteInputStream() = LocalPathLookupsPayload(this).toRemoteInputStream()
 
-fun RemoteInputStream.toLocalPaths() = toLocalPathsPayload().payload
+fun RemoteInputStream.toLocalPathLookups() = toLocalPathLookupsPayload().payload
