+++ app-common-io/src/main/java/eu/darken/sdmse/common/shizuku/service/ShizukuServiceHost.kt
@@ -9,8 +9,8 @@ import eu.darken.rxshell.cmd.RxCmdShell
 import eu.darken.sdmse.common.debug.logging.Logging.Priority.INFO
 import eu.darken.sdmse.common.debug.logging.log
 import eu.darken.sdmse.common.debug.logging.logTag
-import eu.darken.sdmse.common.files.local.root.FileOpsConnection
-import eu.darken.sdmse.common.files.local.root.FileOpsHost
+import eu.darken.sdmse.common.files.local.ipc.FileOpsConnection
+import eu.darken.sdmse.common.files.local.ipc.FileOpsHost
 import eu.darken.sdmse.common.pkgs.pkgops.root.PkgOpsConnection
 import eu.darken.sdmse.common.pkgs.pkgops.root.PkgOpsHost
 import eu.darken.sdmse.common.shell.root.ShellOpsConnection
